tests/check/gst/gstbin.c: It is acceptable to have a refcount of 2 or 3 at this point in the test, because the pipeli...
Original commit message from CVS:
* tests/check/gst/gstbin.c: (GST_START_TEST):
It is acceptable to have a refcount of 2 or 3 at this point in the
test, because the pipeline might be just posting its state_change
message. The next line then waits for that message to appear using
bus_poll, so that should be fine too.